% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.functions.R
\name{alpha.scale}
\alias{alpha.scale}
\title{Generates colours with varying degrees of transparency}
\usage{
alpha.scale(n.cut, col = "blue")
}
\arguments{
\item{col}{Colour to use for shading}

\item{ncut}{A number indicating the number of different counts in the dataset}
}
\description{
Identical to function in \code{MBNMAtime} package
}
