% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.functions.R
\name{calcprob}
\alias{calcprob}
\title{Calculates a matrix of ranking probabilities from a matrix of treatment/agent/class
rankings}
\usage{
calcprob(rank.mat, treats = NULL)
}
\description{
Calculates a matrix of ranking probabilities from a matrix of treatment/agent/class
rankings
}
