% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inconsistency.functions.R
\name{check.indirect.drops}
\alias{check.indirect.drops}
\title{Ensures indirect evidence can be estimated for comparisons identified
within inconsistency.loops}
\usage{
check.indirect.drops(data = data, comp)
}
\description{
Requires repeatedly compiling \code{mbnma.network} objects to identify whether
nodes remain connected by indirect evidence.
}
