% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.functions.R
\name{check.network}
\alias{check.network}
\title{Check if all nodes in the network are connected (identical to function in \code{MBNMAtime})}
\usage{
check.network(g, reference = 1)
}
\description{
Check if all nodes in the network are connected (identical to function in \code{MBNMAtime})
}
