% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose.functions.R
\name{dexp}
\alias{dexp}
\title{Exponential dose-response function}
\usage{
dexp()
}
\value{
An object of \code{class("dosefun")}
}
\description{
Modelled assuming relative effects (\code{"rel"})
}
\details{
\eqn{rate\times{(1-exp(-x))}}

Dose-response parameter arguments:\tabular{ll}{
   \strong{Argument} \tab \strong{Model specification} \cr
   \code{"rel"} \tab Implies that \emph{relative} effects should be pooled for this dose-response parameter separately for each agent in the network. \cr
   \code{"common"} \tab Implies that all studies estimate the same true \emph{absolute} effect for this dose-response parameter across the whole network, \strong{independently of agent}. \cr
   \code{"random"} \tab Implies that all studies estimate separate true \emph{absolute} effects for this dose-response parameter, but that each of these true effects are exchangeable around a true mean effect that is \strong{independent of agent}. This approach allows for modelling of between-study heterogeneity of absolute effects. \cr
   \code{numeric()} \tab Assigned a numeric value, indicating that this dose-response parameter should not be estimated from the data but should be assigned the numeric value determined by the user. This can be useful for fixing specific dose-response parameters (e.g. Hill parameters in Emax functions) to a single value. \cr
}
}
\examples{
dexp()

}
\references{
\insertAllCited
}
