% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.functions.R
\name{disp.obs}
\alias{disp.obs}
\title{Overlays observations as shaded regions on a time-course}
\usage{
disp.obs(g, network, predict, col = "red", max.col.scale = NULL)
}
\arguments{
\item{g}{An object of \code{class("ggplot")}}

\item{network}{An object of class \code{mbnma.network}.}

\item{predict}{An object of class \code{"mbnma.predict"} generated by
\code{predict("mbnma")}}

\item{max.col.scale}{The maximum rgb numeric value to use for the colour scale}
}
\description{
Overlays observations as shaded regions on a time-course
}
