% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose.functions.R
\name{dnonparam}
\alias{dnonparam}
\title{Non-parameteric dose-response functions}
\usage{
dnonparam(direction = "increasing")
}
\arguments{
\item{direction}{Can take either \code{"increasing"} or \code{"decreasing"} to indicate the monotonic direction
of the dose-response relationship}
}
\value{
An object of \code{class("dosefun")}
}
\description{
Used to fit monotonically increasing non-parametric dose-response relationship following
the method of \insertCite{owen2015;textual}{MBNMAdose})
}
\examples{
# Monotonically increasing dose-response
dnonparam(direction="increasing")

# Monotonically decreasing dose-response
dnonparam(direction="decreasing")

}
\references{
\insertAllCited
}
