% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inconsistency.functions.R
\name{drop.comp}
\alias{drop.comp}
\title{Drop treatments from multi-arm (>2) studies for node-splitting}
\usage{
drop.comp(ind.df, drops, comp, start = rbinom(1, 1, 0.5))
}
\description{
Drops arms in a way which preserves connectivity and equally removes
data from each treatment in a nodesplit comparison (so as to maximise precision)
}
