% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.functions.R
\name{E0.validate}
\alias{E0.validate}
\title{Checks the validity of ref.estimate}
\usage{
E0.validate(data.ab, likelihood = NULL)
}
\description{
Ensures \code{E0} takes the correct form to allow for synthesis of network
reference treatment response
}
