% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.functions.R
\name{gen.parameters.to.save}
\alias{gen.parameters.to.save}
\title{Automatically generate parameters to save for a dose-response MBNMA model}
\usage{
gen.parameters.to.save(model.params, model)
}
\arguments{
\item{model.params}{A character or numeric vector containing the names of the
dose-response parameters in the model}

\item{model}{A JAGS model written as a character object}
}
\description{
Identical to \code{gen.parameters.to.save()} in \code{MBNMAtime}
}
