% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.functions.R
\name{genspline}
\alias{genspline}
\title{Generates spline basis matrices for fitting to dose-response function}
\usage{
genspline(
  x,
  spline = "bs",
  df = 1,
  knots = NULL,
  degree = 3,
  max.dose = max(x),
  boundaries = NULL
)
}
\arguments{
\item{x}{A numeric vector indicating all time points available in the dataset}

\item{spline}{Indicates the type of spline function. Can be either a
natural cubic spline (\code{"ns"}), or B-spline (\code{"bs"}).}

\item{df}{degrees of freedom.  One can supply \code{df} rather than
    knots; \code{ns()} then chooses \code{df - 1 - intercept} knots at
    suitably chosen quantiles of \code{x} (which will ignore missing
    values).  The default, \code{df = NULL}, sets the number of
    inner knots as \code{length(knots)}.}

\item{knots}{Indicates the number/location of internal knots. If a single whole number \verb{>=1} is given
it indicates the number of equally-spaced internal knots. Otherwise (a vector, or a non-integer value)
the values are treated as the quantile locations of the knots as a proportion of the maximum dose in the
dataset. For example, if the maximum dose in the dataset is 100mg/d for a particular agent, \code{knots=c(0.1,0.5)}
would indicate knots should be fitted at 10mg/d and 50mg/d.}

\item{degree}{a positive integer giving the degree of the polynomial from which the spline function is composed
(e.g. \code{degree=3} represents a cubic spline).}

\item{max.dose}{A number indicating the maximum dose between which to calculate the spline function.}

\item{boundaries}{A positive numeric vector of length 2 that represents the doses at which to anchor the B-spline or natural
cubic spline basis matrix. This allows data to extend beyond the boundary knots, or for the basis parameters to not depend on \code{x}.
The default (\code{boundaries=NULL}) is the range of \code{x}.}
}
\value{
A spline basis matrix with number of rows equal to \code{length(x)} and the number of columns equal to the number
of coefficients in the spline.
}
\description{
Generates spline basis matrices for fitting to dose-response function
}
\examples{
x <- 0:100

genspline(x)

# Generate a quadratic B-spline with 1 equally spaced internal knot
genspline(x, spline="bs", df=2, degree=2)

# Generate a natural cubic spline with 3 knots at selected quantiles
genspline(x, spline="ns", knots=c(0.1, 0.5, 0.7))

# Generate a piecewise linear spline with 2 equally spaced knots
genspline(x, spline="bs", degree=1, df=3)

}
