% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.functions.R
\name{get.model.vals}
\alias{get.model.vals}
\title{Get MBNMA model values for dose-response parameters}
\usage{
get.model.vals(mbnma)
}
\arguments{
\item{mbnma}{An S3 object of class \code{"mbnma"} generated by running
a dose-response MBNMA model}
}
\description{
Get MBNMA model values for dose-response parameters
}
