% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.functions.R
\name{get.theta.dev}
\alias{get.theta.dev}
\title{Extracts fitted values or deviance contributions into a data.frame with indices}
\usage{
get.theta.dev(mbnma, param = "theta")
}
\description{
Extracts fitted values or deviance contributions into a data.frame with indices
}
