% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{GoutSUA_2wkCFB}
\alias{GoutSUA_2wkCFB}
\title{Studies of treatments for Serum Uric Acid reduction in patients with gout}
\format{
A data frame in long format (one row per arm and study), with 27 rows and 5 variables:
\itemize{
\item \code{studyID} Study identifiers
\item \code{y} Numeric data indicating the mean change from baseline in SUA in a study arm
\item \code{se} Numeric data indicating the standard error for the mean change from baseline in SUA in a study arm
\item \code{agent} Character data indicating the agent to which participants were randomised
\item \code{dose} Numeric data indicating the standardised dose received
}
}
\source{
Pfizer Ltd.
}
\usage{
GoutSUA_2wkCFB
}
\description{
A dataset from a systematic review of interventions for lowering Serum Uric Acid (SUA) concentration in
patients with gout \strong{(not published previously)}. The outcome is continuous, and aggregate data responses
correspond to the mean change from baseline in SUA in mg/dL at 2 weeks follow-up. The dataset includes 10
Randomised-Controlled Trials (RCTs), comparing 5 different agents, and placebo. Data for one agent (RDEA)
arises from an RCT that is not placebo-controlled, and so is not connected to the network directly. In total
there were 19 different treatments (combination of dose and agent).
}
\keyword{datasets}
