% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.functions.R
\name{MBNMA.comparisons}
\alias{MBNMA.comparisons}
\title{Identify unique comparisons within a network}
\usage{
MBNMA.comparisons(data)
}
\arguments{
\item{data}{A data frame containing variables `studyID` and `treatment` (as numeric codes) that
indicate which treatments are used in which studies.}
}
\value{
A data frame of unique comparisons in which each row represents a different comparison.
`t1` and `t2` indicate the treatment codes that make up the comparison. `nr` indicates the number
of times the given comparison is made within the network.

If there is only a single observation for each study within the dataset (i.e. as for standard
network meta-analysis) `nr` will represent the number of studies that compare treatments `t1` and
`t2`.

If there are multiple observations for each study within the dataset (as in time-course MBNMA)
`nr` will represent the number of time points in the dataset in which treatments `t1` and `t2` are
compared.
}
\description{
Identify unique contrasts within a network that make up all the head-to-head comparisons. Repetitions
of the same treatment comparison are grouped together.
}
\examples{
data <- data.frame(studyID=c(1,1,2,2,3,3,4,4,5,5,5),
  treatment=c(1,2,1,3,2,3,3,4,1,2,4)
  )

# Identify unique comparisons within the data
MBNMA.comparisons(data)


# Using the triptans headache dataset
network <- MBNMA.network(HF2PPITT) # Adds treatment identifiers
MBNMA.comparisons(network$data.ab)

}
