% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.functions.R
\name{mbnma.exponential}
\alias{mbnma.exponential}
\title{Run MBNMA model with a exponential dose-response function}
\usage{
mbnma.exponential(
  network,
  lambda = "rel",
  method = "common",
  class.effect = list(),
  UME = FALSE,
  cor = TRUE,
  var.scale = NULL,
  parameters.to.save = NULL,
  pd = "pd.kl",
  parallel = FALSE,
  likelihood = NULL,
  link = NULL,
  priors = NULL,
  arg.params = NULL,
  ...
)
}
\arguments{
\item{network}{An object of class \code{mbnma.network}.}

\item{lambda}{Refers to the rate of growth/decay of the exponential dose-response function.
Can take either \code{"rel"}, \code{"common"}, \code{"random"}, or be assigned a numeric value (see details).}

\item{method}{Can take either \code{"common"} or \code{"random"} to indicate whether relative effects
should be modelled with between-study heterogeneity or not (see details).}

\item{class.effect}{A list of named strings that determines which dose-response
parameters to model with a class effect and what that effect should be
(\code{"common"} or \code{"random"}). Element names should match dose-response parameter names (which will therefore
depend on whether or not a wrapper function has been used for \code{mbnma.run()}).
For example: \code{list("beta.2"="fixed", "beta.3"="random")} when using
\code{mbnma.run()} or \code{list("ed50"="fixed", "hill"="random")} when using \code{mbnma.emax.hill()}.}

\item{UME}{A boolean object to indicate whether to fit an Unrelated Mean Effects model
that does not assume consistency and so can be used to test if the consistency
assumption is valid.}

\item{cor}{A boolean object that indicates whether correlation should be modelled
between relative effect dose-response parameters (\code{TRUE}) or not (\code{FALSE}). This is
automatically set to \code{FALSE} if class effects are modelled or if multiple dose-response
functions are fitted.}

\item{var.scale}{A numeric vector indicating the relative scale of variances between
correlated dose-response parameters when relative effects are modelled on more than
one dose-response parameter and \code{cor=TRUE} (see details). Each element of
the vector refers to the relative scale of each of the dose-response parameters that is
modelled using relative effects.}

\item{parameters.to.save}{A character vector containing names of parameters
to monitor in JAGS}

\item{pd}{Can take either:
\itemize{
\item \code{pv} only pV will be reported (as automatically outputted by \code{R2jags}).
\item \code{plugin} calculates pD by the plug-in
method \insertCite{spiegelhalter2002}{MBNMAdose}. It is faster, but may output negative
non-sensical values, due to skewed deviances that can arise with non-linear models.
\item \code{pd.kl} calculates pD by the Kullback-Leibler divergence \insertCite{plummer2008}{MBNMAdose}. This
will require running the model for additional iterations but is a more robust calculation for the effective
number of parameters in non-linear models.
\item \code{popt} calculates pD using an optimism adjustment which allows for calculation
of the penalized expected deviance \insertCite{plummer2008}{MBNMAdose}
}}

\item{parallel}{A boolean value that indicates whether JAGS should be run in
parallel (\code{TRUE}) or not (\code{FALSE}). If \code{TRUE} then the number of cores to
use is automatically calculated.}

\item{likelihood}{A string indicating the likelihood to use in the model. Can take either \code{"binomial"},
\code{"normal"} or \code{"poisson"}. If left as \code{NULL} the likelihood will be inferred from the data.}

\item{link}{A string indicating the link function to use in the model. Can take any link function
defined within JAGS (e.g. \code{"logit"}, \code{"log"}, \code{"probit"}, \code{"cloglog"}) or be assigned the value \code{"identity"} for
and identity link function. If left as \code{NULL} the link function will be automatically assigned based
on the likelihood.}

\item{priors}{A named list of parameter values (without indices) and
replacement prior distribution values given as strings
\strong{using distributions as specified in JAGS syntax} (see examples).}

\item{arg.params}{Contains a list of arguments sent to \code{mbnma.run()} by dose-response
specific wrapper functions}

\item{...}{Arguments to be sent to R2jags.}
}
\value{
An object of S3 \code{class(c("mbnma", "rjags"))} containing parameter
results from the model. Can be summarized by \code{print()} and can check
traceplots using \code{R2jags::traceplot()} or various functions from the package \code{mcmcplots}.

Nodes that are automatically monitored (if present in the model) have the
following interpretation:\tabular{ll}{
   \strong{Parameters(s)} \tab \strong{Interpretation} \cr
   \code{d.1}, \code{d.2}, \code{d.3}, \code{d.4} \tab The pooled effect for each agent for a given dose-response parameter. These will be estimated by the model for dose-response parameters (\code{beta.1}, \code{beta.2}, \code{beta.3}, \code{beta.4}) specified as \code{"rel"} (e.g.\code{mbnma.run(beta.1="rel")}) \cr
   \code{sd} \tab The between-study SD (heterogeneity) for relative effects, reported if \code{method="random"} \cr
   \code{D.1}, \code{D.2}, \code{D.3}, \code{D.4} \tab The class effect for each class for a given dose-response parameter. These will be estimated by the model if specified in \code{class.effects} for a given dose-response parameter. \cr
   \code{sd.D.1}, \code{sd.D.2}, \code{sd.D.3}, \code{sd.D.4} \tab The within-class SD for different agents within the same class. Will be estimated by the model if any dose-response parameter in \code{class.effect} is set to \code{"random"}. \cr
   \code{beta.1}, \code{beta.2}, \code{beta.3}, \code{beta.4} \tab The absolute value of a given dose-response parameter across the whole network (i.e. does not vary by agent/class). Will be estimated by the model for dose-response parameters (\code{beta.1}, \code{beta.2}, \code{beta.3}, \code{beta.4}) specified as \code{"common"} or \code{"random"}. \cr
   \code{sd.1}, \code{sd.2}, \code{sd.3}, \code{sd.4} \tab Between-study SD (heterogeneity) for absolute dose-response parameters (\code{beta.1}, \code{beta.2}, \code{beta.3}, \code{beta.4}) specified as \code{"random"}. \cr
   \code{totresdev} \tab The residual deviance of the model \cr
   \code{deviance} \tab The deviance of the model \cr
}


If there are errors in the JAGS model code then the object will be a list
consisting of two elements - an error message from JAGS that can help with
debugging and \code{model.arg}, a list of arguments provided to \code{mbnma.run()}
which includes \code{jagscode}, the JAGS code for the model that can help
users identify the source of the error.
}
\description{
Fits a Bayesian model-based network meta-analysis (MBNMA) with a defined
dose-response function. Follows the methods
of \insertCite{mawdsley2016;textual}{MBNMAdose}. This function acts as a wrapper for \code{mbnma.run()} that
uses more clearly defined parameter names.
}
\section{Dose-response parameter arguments}{
\tabular{ll}{
   \strong{Argument} \tab \strong{Model specification} \cr
   \code{"rel"} \tab Implies that \emph{relative} effects should be pooled for this dose-response parameter separately for each agent in the network. \cr
   \code{"common"} \tab Implies that all studies estimate the same true \emph{absolute} effect for this dose-response parameter across the whole network. \cr
   \code{"random"} \tab Implies that all studies estimate separate true \emph{absolute} effects for this dose-response parameter, but that each of these true effects are exchangeable around a true mean effect. This approach allows for modelling of between-study heterogeneity of absolute effects. \cr
   \code{numeric()} \tab Assigned a numeric value, indicating that this dose-response parameter should not be estimated from the data but should be assigned the numeric value determined by the user. This can be useful for fixing specific dose-response parameters (e.g. Hill parameters in Emax functions) to a single value. \cr
}
}

\examples{
\donttest{
# Using the triptans data
tripnet <- mbnma.network(HF2PPITT)

# Fit a exponential dose-response MBNMA with random treatment effects
exponential <- mbnma.exponential(tripnet, lambda="rel", method="random")

# Fit a exponential dose-response MBNMA using a cloglog link function
exponential <- mbnma.exponential(tripnet, lambda="rel", link="cloglog")


####### Priors #######

# Obtain priors from exponential dose-response MBNMA
exponential <- mbnma.exponential(tripnet, lambda="rel", method="random")
print(exponential$model.arg$priors)

# Set new more informative prior distributions
newpriors <- list(sd = "dnorm(0,0.5) T(0,)")

exponential <- mbnma.exponential(tripnet, lambda="rel", method="random",
                   priors=newpriors)


########## Sampler options ##########

# Change the number of MCMC iterations, the number of chains, and the thin
exponential <- mbnma.exponential(tripnet, lambda="rel", method="random",
                   n.iter=5000, n.thin=5, n.chains=4)

# Calculate effective number of parameters via plugin method
exponential <- mbnma.exponential(tripnet, lambda="rel", method="random",
                   pd="plugin")

# Calculate effective number of parameters via penalized expected deviance
exponential <- mbnma.exponential(tripnet, lambda="rel", method="random",
                   pd="popt")


####### Examine MCMC diagnostics (using mcmcplots package) #######

# Density plots
mcmcplots::denplot(exponential)

# Traceplots
mcmcplots::traplot(exponential)

# Caterpillar plots
mcmcplots::caterplot(exponential, "d.lambda")


########## Output ###########

# Print R2jags output and summary
print(exponential)
summary(exponential)

# Plot forest plot of results
plot(exponential)
}

}
\references{
\insertAllCited
}
