% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.functions.R, R/prepare.functions.R
\name{plot.MBNMA.network}
\alias{plot.MBNMA.network}
\alias{MBNMA.network}
\title{Create an MBNMA.network object}
\usage{
\method{plot}{MBNMA.network}(network, layout_in_circle = TRUE,
  edge.scale = 1, label.distance = 0, level = "treatment",
  remove.loops = FALSE, v.color = "connect", v.scale = NULL,
  doselink = NULL, ...)

MBNMA.network(data.ab, description = "Network")
}
\arguments{
\item{network}{An object of class `MBNMA.network`.}

\item{layout_in_circle}{A boolean value indicating whether the network plot
should be shown in a circle or left as the igraph default layout.}

\item{edge.scale}{A number to scale the thickness of connecting lines
(edges). Line thickness is proportional to the number of studies for a
given comparison. Set to `0` to make thickness equal for all comparisons.}

\item{label.distance}{A number scaling the distance of labels from the nodes
to improve readability. The labels will be directly on top of the nodes if
the defauls of `0` is used. Option only applicable if `layout_in_circle` is
set to `TRUE`.}

\item{level}{A string indicating whether nodes/facets should represent `treatment`
or `agent` in the plot. Can be used to examine the expected impact of modelling
dose-response in terms of network connectivity.}

\item{remove.loops}{A boolean value indicating whether to include loops that
indicate comparisons within a node.}

\item{v.color}{Can take either `"connect"` (the default) to indicate that nodes should
only be coloured if they are connected to the network reference treatment (indicates
network connectivity) or `"agent"` to colour nodes by agent.}

\item{v.scale}{A number with which to scale the size of the nodes. If the variable `N`
(to indicate the numbers of participants in each study arm) is included in the
dataset then the size of the nodes will be proportional to the number of participants
within a treatment/agent in the network.}

\item{doselink}{If given an integer value it indicates that connections via the dose-response
relationship with placebo should be plotted. The integer represents the minimum number of doses
from which a dose-response function could be estimated and is equivalent to the number of
parameters in the desired dose-response function plus one. If left as `NULL` (the default), connections
to placebo via dose-response relationships will not be included.}

\item{...}{Options for plotting in `igraph`.}

\item{data.ab}{A data frame of arm-level data in "long" format containing the columns:
* `studyID` Study identifiers
* `dose` Numeric data indicating the dose (must take positive values)
* `agent` Agent identifiers (can be numeric, factor or character)
* `y` Numeric data indicating the aggregate response for a continuous outcome. Required for
continuous data.
* `se` Numeric data indicating the standard error for a given observation. Required for
continuous data.
* `r` Numeric data indicating the number of responders within a study arm. Required for
binomial or poisson data.
* `N` Numeric data indicating the total number of participants within a study arm. Required for
binomial data
* `E` Numeric data indicating the total exposure time for participants within a study arm. Required
for poisson data.
* `class` An optional column indicating a particular class code. Agents with the same identifier
must also have the same class code.}

\item{description}{Optional. Short description of the network.}
}
\value{
An object of `class("igraph")` - any functions from the `igraph` package
can be applied to this object to change its characteristics.

An object of class `MBNMA.network` which is a list containing:
* `description` A short description of the network
* `data.ab` A data frame containing the arm-level network data (treatment identifiers will have
been recoded to a sequential numeric code)
* `agents` A character vector indicating the agent identifiers that correspond to the
new agent codes.
* `treatments` A character vector indicating the treatment identifiers that correspond
to the new treatment codes.
* `classes` A character vector indicating the class identifiers (if included in the original data)
that correspond to the new class codes.
}
\description{
Creates an object of class `MBNMA.network`. Various MBNMA functions can subsequently be applied
to this object.
}
\details{
The S3 method `plot()` on an `MBNMA.network` object generates a
  network plot that shows how different treatments are connected within the
  network via study comparisons. This can be used to identify how direct and
  indirect evidence are informing different treatment comparisons. Depends on
  \code{\link[igraph]{igraph}}.

Agents/classes for arms that have dose = 0 will be relabelled as `"Placebo"`.
Missing values (`NA`) cannot be included in the dataset. Single arm studies cannot
be included.
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: Generate a network plot
}}

\examples{
# Create an MBNMA.network object from the data
network <- MBNMA.network(HF2PPITT)

# Generate a network plot from the data
plot(network)

# Generate a network plot at the agent level that removes loops indicating comparisons
#within a node
plot(network, level="agent", remove.loops=TRUE)

# Generate a network plot at the treatment level that colours nodes by agent
plot(network, v.color="agent", remove.loops=TRUE)

# Generate a network plot that includes connections via the dose-response function
# For a one parameter dose-response function (e.g. exponential)
plot(network, level="treatment", doselink=1, remove.loops=TRUE)

# For a two parameter dose-response function (e.g. Emax)
plot(network, level="treatment", doselink=2, remove.loops=TRUE)


#### Plot a network with no placebo data included ####
# Make data with no placebo
noplac.df <- network$data.ab[network$data.ab$narm>2 & network$data.ab$agent!=1,]
net.noplac <- MBNMA.network(noplac.df)

# Plotting network automatically plots connections to Placebo via dose-response
plot(net.noplac)
# Using the triptans headache dataset
print(HF2PPITT)

# Define network
network <- MBNMA.network(HF2PPITT, description="Example")

# Plot network
plot(network)

}
