% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.functions.R
\name{MBNMA.validate.data}
\alias{MBNMA.validate.data}
\title{Validates that a dataset fulfills requirements for MBNMA}
\usage{
MBNMA.validate.data(data.ab, single.arm = FALSE)
}
\arguments{
\item{data.ab}{A data frame of arm-level data in "long" format containing the columns:
* `studyID` Study identifiers
* `dose` Numeric data indicating the dose (must take positive values)
* `agent` Agent identifiers (can be numeric, factor or character)
* `y` Numeric data indicating the aggregate response for a continuous outcome. Required for
continuous data.
* `se` Numeric data indicating the standard error for a given observation. Required for
continuous data.
* `r` Numeric data indicating the number of responders within a study arm. Required for
binomial or poisson data.
* `N` Numeric data indicating the total number of participants within a study arm. Required for
binomial data
* `E` Numeric data indicating the total exposure time for participants within a study arm. Required
for poisson data.
* `class` An optional column indicating a particular class code. Agents with the same identifier
must also have the same class code.}
}
\value{
An error if checks are not passed. Runs silently if checks are passed
}
\description{
Validates that a dataset fulfills requirements for MBNMA
}
\details{
Checks done within the validation:
* Checks data.ab has required column names
* Checks there are no NAs
* Checks that all SEs are >0 (if variables are included in dataset)
* Checks that all doses are >=0
* Checks that all r and N are positive (if variables are included in dataset)
* Checks that all y, se, r, N and E are numeric
* Checks that class codes are consistent within each agent
* Checks that agent/class names do not contain restricted characters
* Checks that studies have at least two arms (if `single.arm = FALSE`)
* Checks that each study includes at least two treatments
}
