% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.functions.R
\name{neatCrI}
\alias{neatCrI}
\title{Print results as string with credible intervals in brackets}
\usage{
neatCrI(vals, digits = 3)
}
\description{
Print results as string with credible intervals in brackets
}
