% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inconsistency.functions.R, R/plot.functions.R
\name{NMA.nodesplit}
\alias{NMA.nodesplit}
\alias{plot.NMA.nodesplit}
\title{Node-splitting model for testing consistency at the treatment level}
\usage{
NMA.nodesplit(network, likelihood = NULL, link = NULL,
  method = "common", drop.discon = FALSE, comparisons = NULL, ...)

\method{plot}{NMA.nodesplit}(nodesplit, plot.type = NULL, ...)
}
\arguments{
\item{network}{An object of class `MBNMA.network`.}

\item{likelihood}{A string indicating the likelihood to use in the model. Can take either `"binomial"`,
`"normal"` or `"poisson"`. If left as `NULL` the likelihood will be inferred from the data.}

\item{link}{A string indicating the link function to use in the model. Can take any link function
defined within JAGS (`"logit"`, `"log"`, `"probit"`, `"cloglog"`) or be assigned the value `"identity"` for
and identity link function. If left as `NULL` the link function will be automatically assigned based
on the likelihood.}

\item{method}{Can take either `"common"` or `"random"` to indicate whether relative effects
should be modelled with between-study heterogeneity or not (see LINK `Details`).}

\item{drop.discon}{A boolean object that indicates whether to drop treatments
that are disconnected at the treatment level. Default is `TRUE`. If set to `FALSE` then
this could lead to indentification of nodesplit comparisons that are not connected
to the network reference treatment.}

\item{comparisons}{A matrix specifying the comparisons to be split (one row per comparison).
The matrix must have two columns indicating each treatment for each comparison. Values can
either be character (corresponding to the treatment names given in `network`) or
numerical (corresponding to treatment codes within the `network` - note that these
may change if `drop.discon=TRUE`).}

\item{...}{Arguments to be sent to `R2jags`}

\item{nodesplit}{An object of `class("NMA.nodesplit")`}

\item{plot.type}{A character string that can take the value of `"forest"` to plot
only forest plots, `"density"` to plot only density plots, or left as `NULL` (the
default) to plot both types of plot.}
}
\value{
Plots the desired graph(s) and returns an object (or list of object if
`plot.type=NULL`) of `class(c("gg", "ggplot"))`
}
\description{
Splits contributions for a given set of treatment comparisons into direct and
indirect evidence. A discrepancy between the two suggests that the consistency
assumption required for NMA and MBNMA may violated.
}
\details{
The S3 method `plot()` on an `NMA.nodesplit` object generates either
forest plots of posterior medians and 95% credible intervals, or density plots
of posterior densities for direct and indirect evidence.
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: Plot outputs from treatment-level nodesplit models
}}

\examples{
# Using the triptans data
network <- MBNMA.network(HF2PPITT)

split <- NMA.nodesplit(network, likelihood = "binomial", link="logit",
  method="common")


#### To perform nodesplit on selected comparisons ####

# Check for closed loops of treatments with independent evidence sources
loops <- inconsistency.loops(network$data.ab)

split <- NMA.nodesplit(network, likelihood = "binomial", link="logit",
  method="random", comparisons=rbind(c(6,23), c(6,12)))

# Drop treatments that are disconnected from the network in the analysis
split <- NMA.nodesplit(net.noplac, likelihood = "binomial", link="logit",
  method="random", drop.discon=TRUE)

# Plot results
plot(split, plot.type="density") # Plot density plots of posterior densities
plot(split, plot.type="forest") # Plot forest plots of direct and indirect evidence

# Print and summarise results
print(split)
summary(split) # Generate a data frame of summary results
}
