% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.functions.R
\name{print.mbnma.network}
\alias{print.mbnma.network}
\title{Print mbnma.network information to the console}
\usage{
\method{print}{mbnma.network}(x, ...)
}
\arguments{
\item{...}{Arguments to be sent to R2jags.}
}
\description{
Print mbnma.network information to the console
}
