% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.functions.R
\name{print.MBNMA.nodesplit}
\alias{print.MBNMA.nodesplit}
\title{Prints summary results from an MBNMA.nodesplit object}
\usage{
\method{print}{MBNMA.nodesplit}(nodesplit)
}
\description{
Prints summary results from an MBNMA.nodesplit object
}
