% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.functions.R
\name{print.mbnma.predict}
\alias{print.mbnma.predict}
\title{Print summary information from an mbnma.predict object}
\usage{
\method{print}{mbnma.predict}(predict)
}
\arguments{
\item{predict}{An object of class \code{"mbnma.predict"} generated by
\code{predict("mbnma")}}
}
\description{
Print summary information from an mbnma.predict object
}
