% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.functions.R
\name{print.MBNMA.rank}
\alias{print.MBNMA.rank}
\title{Prints summary information about an MBNMA.rank object}
\usage{
\method{print}{MBNMA.rank}(mbnma.rank)
}
\description{
Prints summary information about an MBNMA.rank object
}
