% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodesplit-class.R
\name{print.nodesplit}
\alias{print.nodesplit}
\title{Prints summary results from a nodesplit object}
\usage{
\method{print}{nodesplit}(x, ...)
}
\arguments{
\item{x}{An object of \code{class("nodesplit")}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Prints summary results from a nodesplit object
}
