% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative.array-class.R
\name{print.relative.array}
\alias{print.relative.array}
\title{Print posterior medians (95\% credible intervals) for table of relative effects/mean
differences between treatments/classes}
\usage{
\method{print}{relative.array}(x, digits = 2, ...)
}
\arguments{
\item{x}{An object of class \code{"relative.array"} generated by \code{get.relative()}}

\item{digits}{An integer indicating the number of significant digits to be used.}

\item{...}{further arguments passed to \code{knitr::kable}}
}
\description{
Print posterior medians (95\% credible intervals) for table of relative effects/mean
differences between treatments/classes
}
