% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.functions.R
\name{print.treat.str}
\alias{print.treat.str}
\title{Neatly prints a summary of results}
\usage{
\method{print}{treat.str}(mbnma)
}
\arguments{
\item{mbnma}{An S3 object of class \code{"mbnma"} generated by running
a dose-response MBNMA model}
}
\description{
Neatly prints a summary of results
}
