% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{psoriasis}
\alias{psoriasis}
\title{Studies of biologics for treatment of moderate-to-severe psoriasis}
\format{
A data frame in long format (one row per arm and study), with 81 rows and 9 variables:
\itemize{
\item \code{studyID} Study identifiers
\item \code{agent} Character data indicating the agent to which participants were randomised
\item \code{dose_mg} Numeric data indicating the dose to which participants were randomised in mg
\item \code{freq} Character data indicating the frequency of the dose to which participants were randomised
\item \code{dose} Numeric data indicating the dose in mg/week to which the participants were randomised
\item \code{N} Numeric data indicating the number of participants randomised
\item \code{r75} Numeric data indicating the number of participants who achieved >=75\% improvement in PASI score after 12 weeks
\item \code{r90} Numeric data indicating the number of participants who achieved >=90\% improvement in PASI score after 12 weeks
\item \code{r100} Numeric data indicating the number of participants who achieved 100\% improvement in PASI score after 12 weeks
}
}
\usage{
psoriasis
}
\description{
A dataset from a systematic review of Randomised-Controlled Trials (RCTs) comparing biologics at different doses and placebo
\insertCite{warren2019}{MBNMAdose}. Three different binary outcomes are included, all based on the number of patients experiencing
degrees of improvement on the Psoriasis Area and Severity Index (PASI) measured at 12 weeks follow-up. The dataset
includes 28 Randomised-Controlled Trials (RCTs), comparing 9 different biologics at different doses with placebo.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
