% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.functions.R
\name{rhat.warning}
\alias{rhat.warning}
\title{Print a warning if any monitored parameters have rhat above the cutoff in an MBNMA model}
\usage{
rhat.warning(mbnma, cutoff = 1.2)
}
\description{
Print a warning if any monitored parameters have rhat above the cutoff in an MBNMA model
}
