% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ssi_closure}
\alias{ssi_closure}
\title{Studies of wound closure methods to reduce Surgical Site Infections (SSI)}
\format{
A data frame in long format (one row per arm and study), with 281 rows and 6 variables:
\itemize{
\item \code{studyID} Study identifiers
\item \code{Year} Year of publication
\item \code{n} Numeric data indicating the number of participants randomised
\item \code{r} Numeric data indicating the number of participants who achieved >50\% improvement in depression symptoms
\item \code{trt} Treatment names, given as character data
\item \code{class} Class names, given as character data
}
}
\usage{
ssi_closure
}
\description{
A dataset from an ongoing systematic review examining the efficacy of different wound closure methods to reduce surgical
site infections (SSI). The outcome is binary and represents the number of patients who experienced a SSI. The dataset
includes 129 RCTs comparing 16 different interventions in 6 classes. This dataset is primarily used to illustrate
how \code{MBNMAdose} can be used to perform different types of network meta-analysis without dose-response information.
}
\keyword{datasets}
