% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ssri}
\alias{ssri}
\title{Studies of Selective Serotonin Reuptake Inhibitors (SSRIs) for major depression}
\format{
A data frame in long format (one row per arm and study), with 145 rows and 8 variables:
\itemize{
\item \code{studyID} Study identifiers
\item \code{bias} Risk of bias evaluated on 6 domains
\item \code{age} Mean participant age
\item \code{weeks} Duration of study follow-up
\item \code{agent} Character data indicating the agent to which participants were randomised
\item \code{dose} Numeric data indicating the dose to which participants were randomised in mg
\item \code{N} Numeric data indicating the number of participants randomised
\item \code{r} Numeric data indicating the number of participants who achieved >50\% improvement in depression symptoms
}
}
\usage{
ssri
}
\description{
A dataset from a systematic review examining the efficacy of different doses of SSRI antidepressant drugs and placebo
\insertCite{furukawa2019}{MBNMAdose}. The response to treatment is defined as a 50\% reduction in depressive symptoms after 8 weeks
(4-12 week range) follow-up. The dataset includes 60 RCTs comparing 5 different SSRIs with placebo.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
