% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbnma.network-class.R
\name{summary.mbnma.network}
\alias{summary.mbnma.network}
\title{Print summary mbnma.network information to the console}
\usage{
\method{summary}{mbnma.network}(object, ...)
}
\arguments{
\item{object}{An object of class \code{mbnma.network}.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print summary mbnma.network information to the console
}
