% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.functions.R
\name{summary.MBNMA.nodesplit}
\alias{summary.MBNMA.nodesplit}
\title{Generates a summary data frame for MBNMA.nodesplit objects}
\usage{
\method{summary}{MBNMA.nodesplit}(nodesplit)
}
\description{
Generates a summary data frame for MBNMA.nodesplit objects
}
