% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.functions.R
\name{summary.MBNMA.predict}
\alias{summary.MBNMA.predict}
\title{Produces a summary data frame from an MBNMA.predict object}
\usage{
\method{summary}{MBNMA.predict}(predict)
}
\value{
A data frame containing posterior summary statistics from predicted responses
from a dose-response MBNMA model
}
\description{
Produces a summary data frame from an MBNMA.predict object
}
