% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.functions.R
\name{summary.MBNMA.rank}
\alias{summary.MBNMA.rank}
\title{Generates summary data frames for an MBNMA.rank object}
\usage{
\method{summary}{MBNMA.rank}(mbnma.rank)
}
\value{
A list in which each element represents a parameter that has been ranked
in \code{mbnma.rank} and contains a data frame of summary ranking results.
}
\description{
Generates summary data frames for an MBNMA.rank object
}
