% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodesplit-class.R
\name{summary.nodesplit}
\alias{summary.nodesplit}
\title{Generates a summary data frame for nodesplit objects}
\usage{
\method{summary}{nodesplit}(object, ...)
}
\arguments{
\item{object}{An object of \code{class("nodesplit")}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Generates a summary data frame for nodesplit objects
}
