% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.functions.R
\name{sumrank}
\alias{sumrank}
\title{Generates a summary data frame from a matrix of treatment/agent/class rankings}
\usage{
sumrank(rank.mat)
}
\description{
Generates a summary data frame from a matrix of treatment/agent/class rankings
}
