% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.functions.R
\name{update.mbnma}
\alias{update.mbnma}
\title{Update MBNMA to monitor deviance nodes in the model}
\usage{
\method{update}{mbnma}(mbnma, param = "theta")
}
\arguments{
\item{mbnma}{An S3 object of class `"MBNMA"` generated by running
a dose-response MBNMA model}

\item{param}{Used to indicate which node to monitor in the model. Can be any parameter
in the model code that varies by all arms within all studies. These are some typical
parameters that it might be of interest to monitor, provided they are in the original
model code:
 * `"theta"` for fitted values
 * `"psi"` for fitted values on natural scale (e.g. probabilities)
 * `"dev"` for deviance contributions
 * `"resdev"` for residual deviance contributions
 * `"delta"` for within-study relative effects versus the study reference treatment}
}
\value{
A data frame containing the posterior mean of the updates by arm and study,
with arm and study identifiers.

For MBNMAdose:
* `facet` indicates the agent identifier in the given arm of a study
* `fupdose` indicates the dose in the given arm of a study

For MBNMAtime:
* `facet` indicates the treatment identifier in the given arm of the study
* `fupdose` indicates the follow-up time at the given observation in the given
arm of the study
}
\description{
Useful for obtaining deviance contributions or fitted values
}
\examples{
# Using the triptans data
network <- MBNMA.network(HF2PPITT)

# Fit a dose-response MBNMA, monitoring "psi" and "resdev"
result <- MBNMA.run(network, fun="exponential", beta.1="rel", method="random",
  parameters.to.save=c("psi", "resdev"))

update.mbnma(result, param="theta") # monitor theta

update.mbnma(result, param="rhat") # monitor rhat

update.mbnma(result, param="delta") # monitor delta

}
