% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.functions.R
\name{write.beta.nonparam}
\alias{write.beta.nonparam}
\title{Adds sections of JAGS code for a nonparametric MBNMA model that correspond to beta
parameters}
\usage{
write.beta.nonparam(model, method = "common", fun = "nonparam.up")
}
\description{
Adds sections of JAGS code for a nonparametric MBNMA model that correspond to beta
parameters
}
