% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.functions.R
\name{write.beta}
\alias{write.beta}
\title{Adds sections of JAGS code for an MBNMA model that correspond to beta
parameters}
\usage{
write.beta(model, beta.1, beta.2 = NULL, beta.3 = NULL,
  method = "common", class.effect = list())
}
\arguments{
\item{model}{A character object of JAGS MBNMA model code}

\item{beta.1}{Refers to dose-parameter(s) specified within the dose-response function.
Can take either `"rel"`, `"common"`, `"random"`, or be assigned a numeric value (see details).}

\item{beta.2}{Refers to dose-parameter(s) specified within the dose-response function.
Can take either `"rel"`, `"common"`, `"random"`, or be assigned a numeric value (see details).}

\item{beta.3}{Refers to dose-parameter(s) specified within the dose-response function.
Can take either `"rel"`, `"common"`, `"random"`, or be assigned a numeric value (see details).}

\item{method}{Can take either `"common"` or `"random"` to indicate whether relative effects
should be modelled with between-study heterogeneity or not (see LINK `Details`).}

\item{class.effect}{A list of named strings that determines which dose-response
parameters to model with a class effect and what that effect should be
(`"common"` or `"random"`). For example: `list("beta.2"="fixed", "beta.3"="random")`.}
}
\value{
A character object of JAGS MBNMA model code that includes beta
  parameter components of the model
}
\description{
Adds sections of JAGS code for an MBNMA model that correspond to beta
parameters
}
