% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.functions.R
\name{write.beta.vars}
\alias{write.beta.vars}
\title{Generate objects required for write.beta and write.beta.ref}
\usage{
write.beta.vars()
}
\description{
Generate objects required for write.beta and write.beta.ref
}
