% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.functions.R
\name{write.cov.mat}
\alias{write.cov.mat}
\title{Function for adding covariance matrix for correlation between relative effects}
\usage{
write.cov.mat(model, sufparams, cor = "estimate",
  cor.prior = "wishart", var.scale = NULL)
}
\arguments{
\item{model}{A character object of JAGS MBNMA model code}

\item{sufparams}{A numeric vector of dose-response/time-course parameter suffixes. It
should be the same length as the number of relative effects (i.e. the covariance
matrix size).}

\item{cor}{A boolean object that indicates whether correlation should be modelled
between relative effect dose-response parameters (\code{TRUE}) or not (\code{FALSE}). This is
automatically set to \code{FALSE} if class effects are modelled.}

\item{cor.prior}{NOT CURRENTLY IN USE - indicates the prior distribution to use for the correlation/covariance
between relative effects. Must be kept as \code{"wishart"}}

\item{var.scale}{A numeric vector indicating the relative scale of variances between
correlated dose-response parameters when relative effects are modelled on more than
one dose-response parameter and \code{cor=TRUE} (see details). Each element of
the vector refers to the relative scale of each of the dose-response parameters that is
modelled using relative effects.}
}
\description{
Function for adding covariance matrix for correlation between relative effects
}
