% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.functions.R
\name{write.dose.fun}
\alias{write.dose.fun}
\title{Write dose-response function component of JAGS code for MBNMA dose-response
models}
\usage{
write.dose.fun(fun = "linear", user.fun = NULL, effect = "rel")
}
\arguments{
\item{fun}{is a character specifying a functional form to be assigned to the
dose-response. Options are given in `details`.}

\item{user.fun}{A character specifying any relationship including `dose` and
one/several of: `beta.1`, `beta.2`, `beta.3`.}
}
\value{
A single character string containing JAGS model representing the
  dose-response function component of an MBNMA dose-response model, generated
  based on the arguments passed to the function.
}
\description{
Writes a single line of JAGS code representing the dose-response function
component of an MBNMA dose-response model, returned as a single character
string.
}
\details{
When relative effects are modelled on more than one dose-response parameter and
`cor=TRUE`, correlation between the dose-response parameters is automatically
estimated using a vague Wishart prior. This prior can be made slightly more informative
by specifying the relative scale of variances between the dose-response parameters using
`var.scale`.
}
\examples{
# Write a linear dose-response function
write.dose.fun(fun="linear")

# Write an Emax dose-response function without a Hill parameter
write.dose.fun(fun="emax")

# Write an Emax dose-response function with a Hill parameter
write.dose.fun(fun="emax.hill")

# Write a user-defined dose-response function
doseresp <- "beta.1 + (dose ^ beta.2)"
write.dose.fun(fun="user", user.fun=doseresp)
}
