% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.functions.R
\name{write.E0.synth}
\alias{write.E0.synth}
\title{Write E0 synthesis JAGS model}
\usage{
write.E0.synth(synth = "fixed", likelihood = NULL, link = NULL)
}
\arguments{
\item{synth}{A character object that can take the value \code{"fixed"} or \code{"random"} to
specify the the type of pooling to use for synthesis of \code{E0} if a data frame
has been provided for it. Using \code{"random"} rather
than \code{"fixed"} for \code{synth} will result in wider 95\% CrI for predictions.}
}
\description{
Write E0 synthesis JAGS model
}
