% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.functions.R
\name{write.inserts}
\alias{write.inserts}
\title{Writes insert points for RegEx in MBNMA JAGS code}
\usage{
write.inserts()
}
\value{
A list with named elements containing character strings that match
  points in MBNMA JAGS code. These points can therefore be used to insert
  other lines of JAGS code into the correct section within the overall model
  code.
}
\description{
Writes insert points for RegEx in MBNMA JAGS code
}
\examples{
inserts <- write.inserts()

}
