% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.functions.R
\name{write.remove.loops}
\alias{write.remove.loops}
\title{Removes any loops from MBNMA model JAGS code that do not contain any
expressions}
\usage{
write.remove.loops(model)
}
\arguments{
\item{model}{A character object of JAGS MBNMA model code}
}
\value{
A character object of JAGS MBNMA model code that has had empty loops
  removed from it
}
\description{
Removes any loops from MBNMA model JAGS code that do not contain any
expressions
}
