% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.functions.R
\name{calcprob}
\alias{calcprob}
\title{Calculates a matrix of ranking probabilities from a matrix of treatment/agent/class
rankings}
\usage{
calcprob(rank.mat, treats = NULL)
}
\arguments{
\item{rank.mat}{Numeric matrix of treatment/agent/class rankings}

\item{treats}{A numeric vector of treatment codes for which to calculate ranking probabilities}
}
\description{
Calculates a matrix of ranking probabilities from a matrix of treatment/agent/class
rankings
}
