% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.functions.R
\name{changepd}
\alias{changepd}
\title{Update model fit statistics depending on calculation for pD}
\usage{
changepd(model, jagsdata = NULL, pd = "pv", likelihood = NULL, type = "dose")
}
\arguments{
\item{model}{A model object of class \code{"rjags"}}

\item{jagsdata}{A list object containing data used to estimate \code{model}}

\item{likelihood}{A string indicating the likelihood to use in the model. Can take either \code{"binomial"},
\code{"normal"} or \code{"poisson"}. If left as \code{NULL} the likelihood will be inferred from the data.}

\item{type}{Can take either \code{"dose"} for a dose-response MBNMA or \code{"time"} for a
time-course MBNMA (this accounts for multiple observations within an arm)}
}
\value{
A list containing \code{pd} (effective number of parameters calculated using the method
specified in arguments), \code{deviance} (the posterior median of the total residual deviance)
and \code{dic} (the model DIC)
}
\description{
Update model fit statistics depending on calculation for pD
}
