% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.functions.R
\name{devplot}
\alias{devplot}
\title{Plot deviance contributions from an MBNMA model}
\usage{
devplot(mbnma, plot.type = "scatter", facet = TRUE,
  dev.type = "resdev", ...)
}
\arguments{
\item{mbnma}{An S3 object of class \code{"mbnma"} generated by running
a dose-response MBNMA model}

\item{plot.type}{Deviances can be plotted either as scatter points (\code{"scatter"} - the default)
or as boxplots (\code{"box"})}

\item{facet}{A boolean object that indicates whether or not to facet (by agent for \code{MBNMAdose}
and by treatment for \code{MBNMAtime})}

\item{dev.type}{\emph{STILL IN DEVELOPMENT FOR MBNMAdose!} Deviances to plot - can be either residual
deviances (\code{"resdev"}, the default) or deviances (\code{"dev"})}

\item{...}{Arguments to be sent to \code{ggplot2::geom_point()} or \code{ggplot2::geom_boxplot}}
}
\value{
Generates a plot of deviance contributions and returns a list containing the
plot (as an object of \code{class(c("gg", "ggplot"))}), and a data.frame of posterior mean
deviance/residual deviance contributions for each observation.
}
\description{
Plot deviance contributions from an MBNMA model
}
\details{
Deviances should only be plotted for models that have converged successfully. If deviance
contributions have not been monitored in \code{mbnma$parameters.to.save} then additional
iterations will have to be run to get results for these.

For \code{MBNMAtime}, deviance contributions cannot be calculated for models with a multivariate likelihood (i.e.
those that account for correlation between observations) because the covariance matrix in these
models is treated as unknown (if \code{rho = "estimate"}) and deviance contributions will be correlated.
}
\examples{
###########################
###### For MBNMAdose ######
###########################

# Using the triptans data
network <- mbnma.network(HF2PPITT)

# Run an Emax dose-response MBNMA and predict responses
emax <- mbnma.emax(network, method="random")

# Plot deviances
devplot(emax)

# Plot deviances using boxplots
devplot(emax, plot.type="box")

# Plot deviances on a single plot (not facetted by agent)
devplot(emax, facet=FALSE, plot.type="box")

# A data frame of deviance contributions can be obtained from the object
#returned by `devplot`
devs <- devplot(emax)
head(devs$dev.data)

# Other deviance contributions not currently implemented but in future
#it will be possible to plot them like so
#devplot(emax, dev.type="dev")


###########################
###### For MBNMAtime ######
###########################

}
