% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.functions.R
\name{disp.obs}
\alias{disp.obs}
\title{Overlays observations as shaded regions on a time-course plot}
\usage{
disp.obs(g, network, predict, col = "red", max.col.scale = NULL)
}
\description{
Overlays observations as shaded regions on a time-course plot
}
