% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.functions.R
\name{DR.comparisons}
\alias{DR.comparisons}
\title{Adds placebo comparisons for dose-response relationship}
\usage{
DR.comparisons(data.ab, level = "treatment", doselink = NULL)
}
\description{
Function adds additional rows to a data.frame of comparisons in a network that account
for the relationship between placebo and other agents via the dose-response
relationship.
}
