% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose.functions.R
\name{dspline}
\alias{dspline}
\title{Spline dose-response functions}
\usage{
dspline(type = "bs", knots = NULL, degree = 3, df = NULL, betas = "rel")
}
\arguments{
\item{type}{The type of spline. Can take \code{"bs"} (\href{https://mathworld.wolfram.com/B-Spline.html}{B-spline}),
or \code{"ns"} (\href{https://mathworld.wolfram.com/CubicSpline.html}{natural cubic spline}).
Piecewise linear splines can be fitted using \code{"bs"} with \code{degree=1}.}

\item{knots}{Indicates the number/location of internal knots. If a single whole number \verb{>=1} is given
it indicates the number of equally-spaced internal knots. Otherwise (a vector, or a non-integer value)
the values are treated as the quantile locations of the knots as a proportion of the maximum dose in the
dataset. For example, if the maximum dose in the dataset is 100mg/d for a particular agent, \code{knots=c(0.1,0.5)}
would indicate knots should be fitted at 10mg/d and 50mg/d.}

\item{degree}{a positive integer giving the degree of the polynomial from which the spline function is composed
(e.g. \code{degree=3} represents a cubic spline).}

\item{df}{degrees of freedom.  One can supply \code{df} rather than
    knots; \code{ns()} then chooses \code{df - 1 - intercept} knots at
    suitably chosen quantiles of \code{x} (which will ignore missing
    values).  The default, \code{df = NULL}, sets the number of
    inner knots as \code{length(knots)}.}

\item{betas}{A vector of beta parameters corresponding to each spline coefficient. The length must be equal
to the number of dose-response parameters generated by the spline function. If a single value is given then this
specification will be applied to all beta paramters in the model. Can take \code{"rel"}, \code{"common"}, \code{"random"} or be
assigned a numeric value (see details).}
}
\value{
An object of \code{class("dosefun")}
}
\description{
Used to fit B-splines and natural cubic
splines\insertCite{perperoglu2019}{MBNMAdose}.
}
\section{Dose-response parameters}{
\tabular{ll}{
   \strong{Argument} \tab \strong{Model specification} \cr
   \code{"rel"} \tab Implies that \emph{relative} effects should be pooled for this dose-response parameter separately for each agent in the network. \cr
   \code{"common"} \tab Implies that all agents share the same common effect for this dose-response parameter. \cr
   \code{"random"} \tab Implies that all agents share a similar (exchangeable) effect for this dose-response parameter. This approach allows for modelling of variability between agents. \cr
   \code{numeric()} \tab Assigned a numeric value, indicating that this dose-response parameter should not be estimated from the data but should be assigned the numeric value determined by the user. This can be useful for fixing specific dose-response parameters (e.g. Hill parameters in Emax functions) to a single value. \cr
}


When relative effects are modelled on more than one dose-response parameter,
correlation between them is automatically estimated using a vague inverse-Wishart prior.
This prior can be made slightly more informative by specifying the scale matrix \code{omega}
and by changing the degrees of freedom of the inverse-Wishart prior
using the \code{priors} argument in \code{mbnma.run()}.
}

\references{
\insertAllCited
}
