% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.functions.R
\name{gen.parameters.to.save}
\alias{gen.parameters.to.save}
\title{Automatically generate parameters to save for a dose-response MBNMA model}
\usage{
gen.parameters.to.save(model.params, model)
}
\description{
Identical to \code{gen.parameters.to.save()} in \code{MBNMAtime}
}
