% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.functions.R
\name{get.theta.dev}
\alias{get.theta.dev}
\title{Extracts fitted values or deviance contributions into a data.frame with indices}
\usage{
get.theta.dev(mbnma, param = "theta")
}
\arguments{
\item{mbnma}{An S3 object of class \code{"mbnma"} generated by running
a dose-response MBNMA model}

\item{The}{parameter for which to extract values - can take either \code{"theta"}, \code{"dev"} or \code{"resdev"}}
}
\description{
Extracts fitted values or deviance contributions into a data.frame with indices
}
