% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.functions.R
\name{overlay.split}
\alias{overlay.split}
\title{Overlays results of split (treatment-level) NMA on a plot}
\usage{
overlay.split(g, network, method = "common", likelihood = "binomial",
  link = "logit", ...)
}
\description{
Requires automatically running an NMA
}
