% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.functions.R
\name{plot.MBNMA.predict}
\alias{plot.MBNMA.predict}
\title{Plots predicted responses from a dose-response MBNMA model}
\usage{
\method{plot}{MBNMA.predict}(predict, network, disp.obs = FALSE,
  overlay.split = FALSE, method = "common", agent.labs = NULL,
  scales = "free_x", ...)
}
\arguments{
\item{predict}{An object of class `"MBNMA.predict"` generated by
`predict("MBNMA")`}

\item{network}{An object of class `MBNMA.network`.}

\item{disp.obs}{A boolean object to indicate whether to mark which observed doses
were given in the data on the dose-response curves as shaded regions (`TRUE`) or
not (`FALSE`). If set to `TRUE` the original network object used for the model
*must* be specified in `network`.}

\item{overlay.split}{A boolean object indicating whether to overlay a line
showing the split NMA results on the plot (`TRUE`) or not (`FALSE`). This will
require running of a split NMA model.
For `overlay.split=TRUE` the original network object used for the model
*must* be specified in `network`.}

\item{method}{Indicates the type of split NMA to perform when `overlay.split=TRUE`. Can
take either `"common"` or `"random"`.}

\item{agent.labs}{A character vector of agent labels to display on plots. If
left as `NULL` (the default) the names of agents will be taken from `predict`. The position of
each label corresponds to each element of `predict`. The number of labels must equal
the number of active agents in `predict`. If placebo / dose=0 data is included in the predictions
then the label `"Placebo"` *should not* be included in `agent.labs`. It will not be shown
in the final plot since placebo is the point within each plot at which dose = 0 (rather
than a separate agent).}

\item{scales}{Should scales be fixed (\code{"fixed"}, the default),
free (\code{"free"}), or free in one dimension (\code{"free_x"},
\code{"free_y"})?}

\item{...}{Arguments for `ggplot2`}
}
\description{
Plots predicted responses on the natural scale from a dose-response MBNMA model.
}
\details{
For the S3 method `plot()`, it is advisable to ensure predictions in
  `predict` are estimated using a sufficient number of doses to ensure a smooth
  predicted dose-response curve. If `disp.obs` is set to `TRUE` it is
  advisable to ensure predictions in `predict` are estimated using an even
  sequence of time points to avoid misrepresentation of shaded densities.

  If using `overlay.split`, or `disp.obs` then the original dataset must be specified
  by including the original network object used to estimate the model as the `network`
  argument.
}
\examples{
# Using the triptans data
network <- MBNMA.network(HF2PPITT)

# Run an Emax dose-response MBNMA and predict responses
emax <- MBNMA.emax(network, method="random")
pred <- predict(emax, E0 = 0.5)
plot(pred)

# Display observed doses on the plot (must include `network`)
plot(pred, disp.obs=TRUE, network=network)

# Display split NMA results on the plot (must include `network`)
plot(pred, overlay.split=TRUE, network=network)

# Split NMA results estimated using random treatment effects model
plot(pred, overlay.split=TRUE, network=network, method="random")

# Add agent labels
plot(pred, agent.labs=c("Elet", "Suma", "Frov", "Almo", "Zolmi",
  "Nara", "Riza"))

# These labels will throw an error because "Placebo" is included in agent.labs when
#it will not be plotted as a separate panel
#### ERROR ####
#plot(pred, agent.labs=c("Placebo", "Elet", "Suma", "Frov", "Almo", "Zolmi",
#  "Nara", "Riza"))


# If insufficient predictions are made across dose-response function
# then the plotted responses are less smooth and can be misleading
pred <- predict(emax, E0 = 0.5, n.doses=3)
plot(pred)


}
