% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.functions.R
\name{plot.MBNMA}
\alias{plot.MBNMA}
\title{Forest plot for results from dose-response MBNMA models}
\usage{
\method{plot}{MBNMA}(mbnma, params = NULL, agent.labs = NULL,
  class.labs = NULL)
}
\arguments{
\item{mbnma}{An S3 object of class `"MBNMA"` generated by running
a dose-response MBNMA model}

\item{params}{A character vector of dose-response parameters to plot.
Parameters must be given the same name as monitored nodes in `mbnma` and must be
modelled as relative effects (`"rel"`). Can be set to
`NULL` to include all available dose-response parameters estimated by `mbnma`.}

\item{agent.labs}{A character vector of agent labels (including `"Placebo"` if it
has been included in the original network). If left as `NULL` (the default) then
labels will be used as defined in the data.}

\item{class.labs}{A character vector of class labels if `mbnma` was modelled using class effects
(including `"Placebo"` if it
has been included in the original network). If left as `NULL`
(the default) then labels will be used as defined in the data.}
}
\value{
A forest plot of class `c("gg", "ggplot")` that has separate panels for
different dose-response parameters. Results are plotted on the link scale.
}
\description{
Generates a forest plot for dose-response parameters.
}
\examples{
# Using the triptans data
network <- MBNMA.network(HF2PPITT)

# Run an exponential dose-response MBNMA and generate the forest plot
exponential <- MBNMA.run(network, fun="exponential")
plot(exponential)

# Plot only Emax parameters from an Emax dose-response MBNMA
emax <- MBNMA.emax(network, emax="rel", ed50="rel", method="random")
plot(emax, params=c("d.emax"))


#### Forest plots including class effects ####
# Generate some classes for the data
class.df <- HF2PPITT
class.df$class <- ifelse(df$agent=="placebo", "placebo", "active1")
class.df$class <- ifelse(df$agent=="eletriptan", "active2", df$class)
netclass <- MBNMA.network(class.df)
emax <- MBNMA.emax(netclass, emax="rel", ed50="rel", method="random",
  class.effect=list("ed50"="common"))

# Plot forest plot with different labels for classes
plot(emax, class.labs=c("Placebo", "Other Active", "Eletriptan"))

# Since "Placebo" is included in the network, it must be included in labels
# Failure to do so will cause an error
## ERROR ## plot(emax, class.labs=c("Other Active", "Eletriptan"))

}
