% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.functions.R
\name{print.mbnma.network}
\alias{print.mbnma.network}
\title{Print mbnma.network information to the console}
\usage{
\method{print}{mbnma.network}(network, ...)
}
\arguments{
\item{...}{further arguments passed to or from other methods}

\item{x}{An object of class \code{mbnma.network}.}
}
\description{
Print mbnma.network information to the console
}
