% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbnma.network-class.R
\name{print.mbnma.network}
\alias{print.mbnma.network}
\title{Print mbnma.network information to the console}
\usage{
\method{print}{mbnma.network}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mbnma.network}.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print mbnma.network information to the console
}
