% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.functions.R
\name{print.MBNMA.predict}
\alias{print.MBNMA.predict}
\title{Print summary information from an MBNMA.predict object}
\usage{
\method{print}{MBNMA.predict}(predict)
}
\description{
Print summary information from an MBNMA.predict object
}
