% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.functions.R
\name{print.mbnma.predict}
\alias{print.mbnma.predict}
\title{Print summary information from an mbnma.predict object}
\usage{
\method{print}{mbnma.predict}(x, ...)
}
\arguments{
\item{x}{An object of \code{class("mbnma.rank")} generated by \code{rank.mbnma()}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print summary information from an mbnma.predict object
}
