% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.functions.R
\name{print.mbnma.predict}
\alias{print.mbnma.predict}
\title{Print summary information from an mbnma.predict object}
\usage{
\method{print}{mbnma.predict}(predict)
}
\description{
Print summary information from an mbnma.predict object
}
