% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbnma.rank-class.R
\name{print.mbnma.rank}
\alias{print.mbnma.rank}
\title{Prints summary information about an mbnma.rank object}
\usage{
\method{print}{mbnma.rank}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"mbnma.rank"} generated by \code{rank.mbnma()}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Prints summary information about an mbnma.rank object
}
