% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.functions.R
\name{print.nma.nodesplit}
\alias{print.nma.nodesplit}
\title{Prints summary results from an nma.nodesplit object}
\usage{
\method{print}{nma.nodesplit}(nodesplit)
}
\description{
Prints summary results from an nma.nodesplit object
}
