% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.functions.R
\name{print.nma.nodesplit}
\alias{print.nma.nodesplit}
\title{Prints summary results from an nma.nodesplit object}
\usage{
\method{print}{nma.nodesplit}(x, ...)
}
\arguments{
\item{x}{An object of \code{class("nma.nodesplit")}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Prints summary results from an nma.nodesplit object
}
