% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.functions.R
\name{recode.agent}
\alias{recode.agent}
\title{Assigns agent or class variables numeric identifiers}
\usage{
\method{recode}{agent}(data.ab, level = "agent")
}
\arguments{
\item{level}{Can take either \code{"agent"} or \code{"class"}}
}
\value{
A list containing a data frame with recoded agent/class identifiers and
a character vector of original agent/class names
}
\description{
Assigns agent or class variables numeric identifiers
}
\details{
Also relabels the agent for any arms in which dose = 0 to "Placebo_0"
}
