% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbnma.predict-class.R
\name{summary.mbnma.predict}
\alias{summary.mbnma.predict}
\title{Produces a summary data frame from an mbnma.predict object}
\usage{
\method{summary}{mbnma.predict}(object, ...)
}
\arguments{
\item{object}{An object of \verb{class("mbnma.predict)"} generated by
\code{predict("mbnma")}}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
A data frame containing posterior summary statistics from predicted responses
from a dose-response MBNMA model
}
\description{
Produces a summary data frame from an mbnma.predict object
}
