% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.functions.R
\name{summary.mbnma.predict}
\alias{summary.mbnma.predict}
\title{Produces a summary data frame from an mbnma.predict object}
\usage{
\method{summary}{mbnma.predict}(predict)
}
\arguments{
\item{predict}{An object of class \code{"mbnma.predict"} generated by
\code{predict("mbnma")}}
}
\value{
A data frame containing posterior summary statistics from predicted responses
from a dose-response MBNMA model
}
\description{
Produces a summary data frame from an mbnma.predict object
}
