% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbnma.rank-class.R
\name{summary.mbnma.rank}
\alias{summary.mbnma.rank}
\title{Generates summary data frames for an mbnma.rank object}
\usage{
\method{summary}{mbnma.rank}(object, ...)
}
\arguments{
\item{object}{An object of \code{class("mbnma.rank")} generated by \code{rank.mbnma()}}

\item{...}{additional arguments affecting the summary produced}
}
\value{
A list in which each element represents a parameter that has been ranked
in \code{mbnma.rank} and contains a data frame of summary ranking results.
}
\description{
Generates summary data frames for an mbnma.rank object
}
