% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.functions.R
\name{summary.mbnma}
\alias{summary.mbnma}
\title{Print summary of MBNMA results to the console}
\usage{
\method{summary}{mbnma}(x, ...)
}
\arguments{
\item{x}{An S3 object of class \code{"mbnma"} generated by running
a dose-response MBNMA model}

\item{...}{additional arguments affecting the summary produced}
}
\description{
Print summary of MBNMA results to the console
}
