% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbnma-class.R
\name{summary.mbnma}
\alias{summary.mbnma}
\title{Print summary of MBNMA results to the console}
\usage{
\method{summary}{mbnma}(object, digits = 4, ...)
}
\arguments{
\item{object}{An S3 object of class \code{"mbnma"} generated by running
a dose-response MBNMA model}

\item{digits}{The maximum number of digits for numeric columns}

\item{...}{additional arguments affecting the summary produced}
}
\description{
Print summary of MBNMA results to the console
}
