% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.nodesplit-class.R
\name{summary.nma.nodesplit}
\alias{summary.nma.nodesplit}
\title{Generates a summary data frame for nma.nodesplit objects}
\usage{
\method{summary}{nma.nodesplit}(object, ...)
}
\arguments{
\item{object}{An object of \code{class("nma.nodesplit")}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Generates a summary data frame for nma.nodesplit objects
}
