% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.functions.R
\name{summary.nma.nodesplit}
\alias{summary.nma.nodesplit}
\title{Generates a summary data frame for mbnma.nodesplit objects}
\usage{
\method{summary}{nma.nodesplit}(nodesplit)
}
\arguments{
\item{nodesplit}{An object of \code{class("nma.nodesplit")}}
}
\description{
Generates a summary data frame for mbnma.nodesplit objects
}
