% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.functions.R
\name{summary.nma.nodesplit}
\alias{summary.nma.nodesplit}
\title{Generates a summary data frame for mbnma.nodesplit objects}
\usage{
\method{summary}{nma.nodesplit}(x, ...)
}
\arguments{
\item{x}{An object of \code{class("nma.nodesplit")}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Generates a summary data frame for mbnma.nodesplit objects
}
