% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.functions.R
\name{write.cor}
\alias{write.cor}
\title{Adds correlation between dose-response relative effects}
\usage{
write.cor(model, cor = TRUE, cor.prior = "wishart", var.scale = NULL,
  beta.1 = "rel", beta.2 = NULL, beta.3 = NULL, method = "random",
  class.effect = list())
}
\arguments{
\item{model}{A character object of JAGS MBNMA model code}

\item{cor}{A boolean object that indicates whether correlation should be modelled
between relative effect dose-response parameters (`TRUE`) or not (`FALSE`). This is
automatically set to `FALSE` if class effects are modelled.}

\item{var.scale}{A numeric vector indicating the relative scale of variances between
correlated dose-response parameters when relative effects are modelled on more than
one dose-response parameter and `cor=TRUE` (see `Details` LINK). Each element of
the vector refers to the relative scale of each of the dose-response parameters that is
modelled using relative effects.}

\item{beta.1}{Refers to dose-parameter(s) specified within the dose-response function.
Can take either `"rel"`, `"common"`, `"random"`, or be assigned a numeric value (see details).}

\item{beta.2}{Refers to dose-parameter(s) specified within the dose-response function.
Can take either `"rel"`, `"common"`, `"random"`, or be assigned a numeric value (see details).}

\item{beta.3}{Refers to dose-parameter(s) specified within the dose-response function.
Can take either `"rel"`, `"common"`, `"random"`, or be assigned a numeric value (see details).}

\item{method}{Can take either `"common"` or `"random"` to indicate whether relative effects
should be modelled with between-study heterogeneity or not (see LINK `Details`).}

\item{class.effect}{A list of named strings that determines which dose-response
parameters to model with a class effect and what that effect should be
(`"common"` or `"random"`). For example: `list("beta.2"="fixed", "beta.3"="random")`.}
}
\description{
This uses a Wishart prior as default for modelling the correlation
}
