% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.functions.R
\name{write.cov.mat}
\alias{write.cov.mat}
\title{Function for adding covariance matrix for correlation between relative effects}
\usage{
write.cov.mat(model, sufparams, cor = "estimate",
  cor.prior = "wishart", var.scale = NULL)
}
\arguments{
\item{sufparams}{A numeric vector of dose-response/time-course parameter suffixes. It
should be the same length as the number of relative effects (i.e. the covariance
matrix size).}
}
\description{
Function for adding covariance matrix for correlation between relative effects
}
