% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.functions.R
\name{write.NMA}
\alias{write.NMA}
\title{Write JAGS code for split NMA}
\usage{
write.NMA(method = "common", likelihood = "binomial", link = "logit",
  UME = FALSE)
}
\description{
Write JAGS code for split NMA
}
